+++ ports/mimxrt/cyw43_configport.h
@@ -28,12 +28,8 @@
 #define MICROPY_INCLUDED_MIMXRT_CYW43_CONFIGPORT_H
 
 // The board-level config will be included here, so it can set some CYW43 values.
-#include "py/mpconfig.h"
-#include "py/mperrno.h"
-#include "py/mphal.h"
-#include "extmod/modnetwork.h"
 #include "extmod/mpbthci.h"
-#include "pendsv.h"
+#include "extmod/cyw43_config_common.h"
 #include "sdio.h"
 
 #define CYW43_USE_SPI                   (0)
@@ -61,54 +57,14 @@
 #define CYW43_BT_UART_BAUDRATE_DOWNLOAD_FIRMWARE MICROPY_HW_BLE_UART_BAUDRATE_DOWNLOAD_FIRMWARE
 #endif
 
-#define CYW43_IOCTL_TIMEOUT_US          (1000000)
-#define CYW43_SLEEP_MAX                 (50)
-#define CYW43_NETUTILS                  (1)
 #define CYW43_CLEAR_SDIO_INT            (1)
-
-#define CYW43_EPERM                     MP_EPERM // Operation not permitted
-#define CYW43_EIO                       MP_EIO // I/O error
-#define CYW43_EINVAL                    MP_EINVAL // Invalid argument
-#define CYW43_ETIMEDOUT                 MP_ETIMEDOUT // Connection timed out
-
-#define CYW43_THREAD_ENTER              MICROPY_PY_LWIP_ENTER
-#define CYW43_THREAD_EXIT               MICROPY_PY_LWIP_EXIT
-#define CYW43_THREAD_LOCK_CHECK
-
-#define CYW43_HOST_NAME                 mod_network_hostname_data
 
 #define CYW43_SDPCM_SEND_COMMON_WAIT    __WFI();
 #define CYW43_DO_IOCTL_WAIT             __WFI();
 
-#define CYW43_ARRAY_SIZE(a)             MP_ARRAY_SIZE(a)
-
-#define CYW43_HAL_PIN_MODE_INPUT        MP_HAL_PIN_MODE_INPUT
-#define CYW43_HAL_PIN_MODE_OUTPUT       MP_HAL_PIN_MODE_OUTPUT
-#define CYW43_HAL_PIN_PULL_NONE         MP_HAL_PIN_PULL_NONE
-#define CYW43_HAL_PIN_PULL_UP           MP_HAL_PIN_PULL_UP
-#define CYW43_HAL_PIN_PULL_DOWN         MP_HAL_PIN_PULL_DOWN
-
-#define CYW43_HAL_MAC_WLAN0             MP_HAL_MAC_WLAN0
-#define CYW43_HAL_MAC_BDADDR            MP_HAL_MAC_BDADDR
-
-#define cyw43_hal_ticks_us              mp_hal_ticks_us
-#define cyw43_hal_ticks_ms              mp_hal_ticks_ms
-
-#define cyw43_hal_pin_obj_t             mp_hal_pin_obj_t
-#define cyw43_hal_pin_read              mp_hal_pin_read
-#define cyw43_hal_pin_low               mp_hal_pin_low
-#define cyw43_hal_pin_high              mp_hal_pin_high
-
-#define cyw43_hal_get_mac               mp_hal_get_mac
-#define cyw43_hal_get_mac_ascii         mp_hal_get_mac_ascii
-#define cyw43_hal_generate_laa_mac      mp_hal_generate_laa_mac
-
 #define cyw43_hal_uart_set_baudrate     mp_bluetooth_hci_uart_set_baudrate
 #define cyw43_hal_uart_write            mp_bluetooth_hci_uart_write
 #define cyw43_hal_uart_readchar         mp_bluetooth_hci_uart_readchar
-
-#define cyw43_delay_us                  mp_hal_delay_us
-#define cyw43_delay_ms                  mp_hal_delay_ms
 
 #define cyw43_bluetooth_controller_init     mp_bluetooth_hci_controller_init
 #define cyw43_bluetooth_controller_deinit   mp_bluetooth_hci_controller_deinit
@@ -135,7 +91,7 @@
 #define CYW43_PIN_RFSW_VDD              pin_WL_RFSW_VDD
 #endif
 
-#define cyw43_schedule_internal_poll_dispatch(func) pendsv_schedule_dispatch(PENDSV_DISPATCH_CYW43, func)
+#undef cyw43_hal_pin_config // mp_hal_pin_config not yet implemented on this port
 
 static inline void cyw43_hal_pin_config(cyw43_hal_pin_obj_t pin, uint8_t mode, uint8_t pull, uint8_t alt) {
     machine_pin_set_mode(pin, mode);
@@ -174,7 +130,5 @@ static inline int cyw43_sdio_transfer(uint32_t cmd, uint32_t arg, uint32_t *resp
 static inline int cyw43_sdio_transfer_cmd53(bool write, uint32_t block_size, uint32_t arg, size_t len, uint8_t *buf) {
     return sdio_transfer_cmd53(write, block_size, arg, len, buf);
 }
-
-#define CYW43_EVENT_POLL_HOOK MICROPY_EVENT_POLL_HOOK
 
 #endif // MICROPY_INCLUDED_MIMXRT_CYW43_CONFIGPORT_H
